requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
SQLAlchemy==1.4.39
python-jose==3.3.0
passlib==1.7.4
pydantic~=1.9.1
Authlib==1.0.1
httpx==0.23.0
oauthlib~=3.2.0
alembic~=1.8.0